move toggles to own components, rename for clarity
